Refactor setup
